+++ source/Common.h
@@ -40,7 +40,7 @@ typedef std::vector<uint64_t> UInt64Vec;
  * (Only exact matches are assumed to be compatible, that's why this can differ from the program
  * version.)
  */
-#define HTTP_PROTOCOLVERSION	"2.0.0"
+#define HTTP_PROTOCOLVERSION	"2.0.1"
 
 
 /**